Gateway packaging rename
