--- components/apimgt/org.wso2.carbon.apimgt.ballerina.gateway.annotations/src/main/java/org/wso2/carbon/apimgt/ballerina/gateway/annotations/models/VersionModel.java
@@ -1,45 +1,0 @@
-/*
- * Copyright (c) WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.carbon.apimgt.ballerina.gateway.annotations.models;
-
-public class VersionModel {
-
-    private static VersionModel instance;
-    private String apiVersion;
-
-    private VersionModel() {
-    }
-
-    public static VersionModel getInstance() {
-        synchronized (VersionModel.class) {
-            if (instance == null) {
-                instance = new VersionModel();
-            }
-        }
-        return instance;
-    }
-
-    public String getApiVersion() {
-        return apiVersion;
-    }
-
-    public void setApiVersion(String apiVersion) {
-        this.apiVersion = apiVersion;
-    }
-}
