--- components/apimgt/org.wso2.carbon.apimgt.ballerina.gateway.annotations/src/main/resources/META-INF/services/org.ballerinalang.compiler.plugins.CompilerPlugin
@@ -1,1 +1,0 @@
-org.wso2.carbon.apimgt.ballerina.gateway.annotations.GatewayAnnotationPlugin
