--- components/apimgt/org.wso2.carbon.apimgt.ballerina.gateway/src/main/ballerina/Ballerina.toml
@@ -1,3 +1,0 @@
-[project]
-# Name of the package
-org-name = "wso2"
