+++ components/apimgt/org.wso2.carbon.apimgt.ballerina.gateway/src/main/ballerina/gateway/global_throttle_stream.bal
