+++ components/apimgt/org.wso2.carbon.apimgt.ballerina.gateway/src/main/resources/META-INF/services/org.ballerinalang.compiler.plugins.CompilerPlugin
@@ -1,0 +1,1 @@
+org.ballerina.gateway.GatewayAnnotationPlugin
